search_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
supabase==1.0.3
pydantic==2.4.2
httpx==0.25.1
python-multipart==0.0.6
langchain==0.0.350
langchain-core==0.0.10
langchain-community==0.0.10
wandb==0.16.1 